Moved BatchProcessLogFrame into org.janelia.mipav
